+++ lab1/check.h
@@ -1,7 +1,7 @@
 #ifndef CHECK_H
 #define CHECK_H
 
-void CheckInput(double x, int n);
+void CheckInput(double x, long long int n);
 void CheckArgc(int argc);
 
 #endif
