+++ lab1/logarithm.h
@@ -1,6 +1,6 @@
 #ifndef LOGARITHM_H
 #define LOGARITHM_H
 
-double CountLogarithm(double x, int n);
+double CountLogarithm(double x, long long int n);
 
 #endif
